requirements.txt
langchain
openai
python-dotenv
litellm
bs4